AiEngine/requirements.txt
fastapi
numpy
matplotlib
Pillow
uvicorn
python-multipart
jsonschema
requests
PyYAML
ultralytics
tensorflow
opencv-python
